# openapi/v1.yaml
...
openapi: "3.0.0"
info:
  version: "1.0.0-alpha1"
  description: "API definition for Gopad, Etherpad for markdown with Go"
  title: "Gopad OpenAPI"
  contact:
    name: "Gopad"
    email: "gopad@webhippie.de"
    url: "https://gopad.eu"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://try.gopad.eu/api/v1"
security:
  - Header: []
  - Bearer: []
  - Basic: []
paths:
  /auth/{provider}/request:
    get:
      summary: "Request the redirect to defined provider"
      operationId: "RequestProvider"
      tags:
        - "auth"
      parameters:
        - $ref: "#/components/parameters/AuthProviderParam"
      security: []
      responses:
        "308":
          description: "Redirect to the provider"
          content:
            text/html:
              schema:
                type: "string"
        "404":
          description: "Provider not found"
          content:
            text/html:
              schema:
                type: "string"
        "500":
          description: "Internal server error"
          content:
            text/html:
              schema:
                type: "string"
  /auth/{provider}/callback:
    get:
      summary: "Callback to parse the defined provider"
      operationId: "CallbackProvider"
      tags:
        - "auth"
      parameters:
        - $ref: "#/components/parameters/AuthProviderParam"
        - $ref: "#/components/parameters/AuthStateParam"
        - $ref: "#/components/parameters/AuthCodeParam"
      security: []
      responses:
        "308":
          description: "Generated expiring token"
          content:
            text/html:
              schema:
                type: "string"
        "412":
          description: "Failed to initialize provider"
          content:
            text/html:
              schema:
                type: "string"
        "404":
          description: "Provider not found"
          content:
            text/html:
              schema:
                type: "string"
        "500":
          description: "Internal server error"
          content:
            text/html:
              schema:
                type: "string"
...
  /auth/redirect:
    post:
      summary: "Retrieve real token after redirect"
      operationId: "RedirectAuth"
      tags:
        - "auth"
      requestBody:
        $ref: "#/components/requestBodies/RedirectAuthBody"
      security: []
      responses:
        "200":
          $ref: "#/components/responses/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/InvalidTokenError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/login:
    post:
      summary: "Authenticate an user by credentials"
      operationId: "LoginAuth"
      tags:
        - "auth"
      requestBody:
        $ref: "#/components/requestBodies/LoginAuthBody"
      security: []
      responses:
        "200":
          $ref: "#/components/responses/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/BadCredentialsError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/refresh:
    get:
      summary: "Refresh an auth token before it expires"
      operationId: "RefreshAuth"
      tags:
        - "auth"
      responses:
        "200":
          $ref: "#/components/responses/RefreshResponse"
        "401":
          $ref: "#/components/responses/InvalidTokenError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/verify:
    get:
      summary: "Verify validity for an authentication token"
      operationId: "VerifyAuth"
      tags:
        - "auth"
      responses:
        "200":
          $ref: "#/components/responses/VerifyResponse"
        "401":
          $ref: "#/components/responses/InvalidTokenError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    Header:
      type: apiKey
      in: header
      name: X-API-Key
    Bearer:
      type: http
      scheme: bearer
    Basic:
      type: http
      scheme: basic
  parameters:
...
    AuthProviderParam:
      in: "path"
      name: "provider"
      description: "An identifier for the auth provider"
      schema:
        type: "string"
      required: true
      x-example: "github"
    AuthStateParam:
      name: "state"
      in: "query"
      required: false
      schema:
        type: "string"
      description: "Auth state"
    AuthCodeParam:
      name: "code"
      in: "query"
      required: false
      schema:
        type: "string"
      description: "Auth code"
...
  requestBodies:
    RedirectAuthBody:
      description: "The redirect token to authenticate"
      required: true
      content:
        application/json:
          schema:
            type: "object"
            required:
              - "token"
            properties:
              token:
                type: "string"
    LoginAuthBody:
      description: "The credentials to authenticate"
      required: true
      content:
        application/json:
          schema:
            type: "object"
            required:
              - "username"
              - "password"
            properties:
              username:
                type: "string"
              password:
                type: "string"
                format: "password"
...
  responses:
    BadRequestError:
      description: "Failed to parse request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
...
    InternalServerError:
      description: "Some internal server error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
...
    InvalidTokenError:
      description: "Failed to generate or validate token"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
    BadCredentialsError:
      description: "Unauthorized with wrong credentials"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
...
    VerifyResponse:
      description: "Metadata of the auth token"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthVerify"
    LoginResponse:
      description: "Generated token with expire date"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthToken"
    RefreshResponse:
      description: "Refreshed token with expire date"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthToken"
    TokenResponse:
      description: "Generated token never expiring"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthToken"
...
  schemas:
    AuthToken:
      type: "object"
      required:
        - "token"
      properties:
        token:
          type: "string"
          readOnly: true
        expires_at:
          type: "string"
          format: "date-time"
          x-omitempty: true
          x-nullable: true
          readOnly: true
    AuthVerify:
      type: "object"
      required:
        - "username"
      properties:
        username:
          type: "string"
          readOnly: true
        created_at:
          type: "string"
          format: "date-time"
          x-omitempty: true
          x-nullable: true
          readOnly: true
...
    Notification:
      title: "Notification"
      description: "Generic response for errors and validations"
      type: "object"
      properties:
        status:
          type: "integer"
          x-omitempty: true
        message:
          type: "string"
          x-omitempty: true
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/Validation"
    Validation:
      title: "Validation"
      description: "General structure to show validation errors"
      type: "object"
      properties:
        field:
          type: "string"
        message:
          type: "string"
...